refactor: notifications lazy loading
